canvas image render fix
