+++ lib/pdfjs-utils/domfacade.js
@@ -2,25 +2,64 @@
 
 const domstubs = require('pdfjs-dist/lib/examples/node/domstubs');
 const stackTrace = require('stack-trace');
-const { ImageData }  = require("canvas");
+const assert = require('assert');
 
 const { JSDOM }  = require("jsdom");
-const utils = require("jsdom/lib/jsdom/living/generated/utils.js");
+const utils = require("jsdom/lib/jsdom/living/generated/utils");
 const impl = utils.implSymbol;
-const HTMLElement = require("jsdom/lib/jsdom/living/generated/HTMLElement.js");
-const HTMLCanvasElement = require("jsdom/lib/jsdom/living/generated/HTMLCanvasElement.js");
+const HTMLElement = require("jsdom/lib/jsdom/living/generated/HTMLElement");
+const HTMLCanvasElement = require("jsdom/lib/jsdom/living/generated/HTMLCanvasElement");
+const Canvas  = require("jsdom/lib/jsdom/utils").Canvas;
 
 Object.defineProperty(HTMLCanvasElement.expose.Window.HTMLCanvasElement.prototype, "_canvas", {
 	get() {
-		return this[impl]["_canvas"];
+		return this[impl]._getCanvas();
 	},
 	enumerable: true,
 	configurable: true
 });
 
-exports.Image = domstubs.Image;
+/**
+ * A readable stream which offers a stream representing the serialization of a
+ * given DOM element (as defined by domstubs.js).
+ */
+class NodeCanvasFactory
+{
+	create(width, height) {
+		assert(width > 0 && height > 0, 'Invalid canvas size');
+		let canvas = new Canvas(width, height);
+		let context = canvas.getContext('2d');
+		return {
+			canvas: canvas,
+			context: context,
+		};
+	}
+
+	reset(canvasAndContext, width, height) {
+		assert(canvasAndContext.canvas, 'Canvas is not specified');
+		assert(width > 0 && height > 0, 'Invalid canvas size');
+		canvasAndContext.canvas.width = width;
+		canvasAndContext.canvas.height = height;
+	}
+
+	destroy(canvasAndContext) {
+		assert(canvasAndContext.canvas, 'Canvas is not specified');
+
+		// Zeroing the width and height cause Firefox to release graphics
+		// resources immediately, which can greatly reduce memory consumption.
+		canvasAndContext.canvas.width = 0;
+		canvasAndContext.canvas.height = 0;
+		canvasAndContext.canvas = null;
+		canvasAndContext.context = null;
+	}
+}
+
+exports.NodeCanvasFactory = NodeCanvasFactory;
+exports.Canvas = Canvas;
+
 exports.HTMLElement = HTMLElement.expose.Window.HTMLElement;
-exports.ImageData = ImageData;
+exports.Image = Canvas.Image;
+exports.ImageData = Canvas.ImageData;
 exports.jsDomFactory = function () {
 	return new JSDOM(`<html>
 <head>
