+++ lib/renderer/CanvasRenderer.js
@@ -1,47 +1,12 @@
 'use strict';
 
-const assert = require('assert');
-const Canvas = require('canvas');
 const AbstractRenderer = require('./AbstractRenderer');
 const JsonWriter = require('./writer/JsonWriter');
 const PngWriter = require('./writer/PngWriter');
 const HtmlWriter = require('./writer/HtmlWriter');
 const TextWriter = require('./writer/TextWriter');
 
-/**
- * A readable stream which offers a stream representing the serialization of a
- * given DOM element (as defined by domstubs.js).
- */
-class NodeCanvasFactory
-{
-	create(width, height) {
-		assert(width > 0 && height > 0, 'Invalid canvas size');
-		let canvas = new Canvas(width, height);
-		let context = canvas.getContext('2d');
-		return {
-			canvas: canvas,
-			context: context,
-		};
-	}
-
-	reset(canvasAndContext, width, height) {
-		assert(canvasAndContext.canvas, 'Canvas is not specified');
-		assert(width > 0 && height > 0, 'Invalid canvas size');
-		canvasAndContext.canvas.width = width;
-		canvasAndContext.canvas.height = height;
-	}
-
-	destroy(canvasAndContext) {
-		assert(canvasAndContext.canvas, 'Canvas is not specified');
-
-		// Zeroing the width and height cause Firefox to release graphics
-		// resources immediately, which can greatly reduce memory consumption.
-		canvasAndContext.canvas.width = 0;
-		canvasAndContext.canvas.height = 0;
-		canvasAndContext.canvas = null;
-		canvasAndContext.context = null;
-	}
-}
+const { NodeCanvasFactory }  = require("../pdfjs-utils/domfacade");
 
 class CanvasRenderer extends AbstractRenderer
 {
